pyproject.toml
[project]
name = "tv-systools"
version = "0.5.5"
description = "Collection of system-related small tools"
authors = [{ name = "Thorsten Vitt", email = "thorsten.vitt@uni-wuerzburg.de" }]
dependencies = [
  "rich>=13.7.0",
  "python-apt @ git+https://salsa.debian.org/apt-team/python-apt.git",
  "python-dateutil>=2.8.2",
  "ruyaml>=0.91.0",
  "notify2>=0.3.1",
  "pyudev>=0.24.1",
  "typer>=0.9.0",
  "tqdm>=4.66.1",
  "questionary>=2.0.1",
  "platformdirs>=4.2.2",
  "pyxdg>=0.28",
  "prompt-toolkit",
  "pzp>=0.0.25",
  "humanize>=4.11.0",
  "toml>=0.10.2",
  "python-magic>=0.4.27",
  "lxml>=5.3.1",
  "cyclopts>=4.2.1",
  "httpx>=0.28.1",
  "tomlkit>=0.13.3",
  "python-jsonpath>=1.3.1",
  "logproc",
]
requires-python = ">=3.13"
readme = "README.md"
license = { text = "MIT" }

[project.scripts]
apt-describe-update = "tv_systools.apt_describe_update:main"
fetch-and-link = "tv_systools.fetch_and_link:main"
wacom-config = "tv_systools.wacom_config:main"
glyphinfo = "tv_systools.glyphinfo:app"
debug-gnome-extensions = "tv_systools.debug_gnome_extensions:main"
nco = "tv_systools.nco:app"
j = "tv_systools.logseq_j:main"
clean-sync-conflicts = "tv_systools.clean_sync_conflicts:app"
public-link = "tv_systools.public_link:app"
monitor-switch = "tv_systools.monitor_switch:main"
json2toml = "tv_systools.jsontotoml:app"
aps = "tv_systools.apt_search:app"
wh = "tv_systools.superwhich:main"
extract-mw-backup = "tv_systools.extract_mw_backup:app"
pipxtouvtools = "tv_systools.pipxtouvtools:get_pipx_specs"
linked-assets = "tv_systools.linked_assets:app"
nstotop = "tv_systools.nstotop:app"
ta = "tv_systools.tasmota:app"
git-fix-remotes = "tv_systools.fix_github_remote:app"
pdfshrink = "tv_systools.pdfshrink:app"

[project.optional-dependencies]
gui = [
  "pygobject>=3.50.0",
  "readchar>=4.0.5",
  "pydbus>=0.6.0",
  "systemd-python>=235",
]


[tool.pdm]
distribution = true

[tool.pdm.build.wheel-data]
scripts = ["scripts/*"]
data = ["data/**"]

[tool.uv.sources]
logproc = { path = "../logproc" }

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[dependency-groups]
dev = [
  "better-exceptions>=0.3.3",
  "bpython>=0.24",
  "ipykernel>=6.29.5",
  "ipython>=8.18.0",
  "lxml-stubs>=0.5.1",
  "prompt-toolkit>3.0.40",
]
